refactor: ベンチマークファイルをテスト対象の処理単位で再整理
旧・新パターンで分けていたファイルを処理単位で再整理。
旧・新を同一ファイル内で並べることで比較しやすくした。

- bench_strings_test.go: BenchmarkOld + BenchmarkNew
- bench_file_test.go: BenchmarkOldWithSetup + BenchmarkNewWithSetup
